requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-security
flask-script